create mod for ast
