+++ crates/syntax/src/abstract_syntax_tree/mod.rs
@@ -1,0 +1,35 @@
+#[macro_export]
+macro_rules! ast_node {
+    ($ast_name: ident, $kind: expr) => {
+        #[derive(Debug, Clone)]
+        pub struct $ast_name {
+            syntax: SyntaxNode,
+        }
+        impl AstNode for $ast_name {
+            type Language = CircomLanguage;
+            fn can_cast(token_kind: TokenKind) -> bool {
+                token_kind == $kind
+            }
+
+            fn cast(syntax: SyntaxNode) -> Option<Self>
+            where
+                Self: Sized,
+            {
+                if Self::can_cast(syntax.kind().into()) {
+                    return Some(Self { syntax });
+                }
+                None
+            }
+
+            fn syntax(&self) -> &SyntaxNode {
+                &self.syntax
+            }
+        }
+    };
+}
+
+pub mod ast;
+pub mod template;
+
+pub use ast::*;
+pub use template::*;
