+++ crates/syntax/src/lib.rs
@@ -1,3 +1,4 @@
-pub mod abstract_syntax_tree;
 pub mod syntax;
 pub mod syntax_node;
+
+pub mod abstract_syntax_tree;
